Fix module packaging (#24)
* Removed index.ts files from library

It looks like Angular metadata resolver does not process them properly
and resolves symbols to undefined if they are coming from re-export. I
don't completely understand how it works, but after debugging this I
also checked https://github.com/angular/angular/tree/master/packages and
https://github.com/ng-bootstrap/ng-bootstrap/tree/master/src and none
seem to have index.ts files besides the main one.

* Replaced @Component annotation with @Injectable() on the abstract class

If @Component is present compiler attempts to handle this class as a real
component and fails because it was not added to the NgModule. What is
really needed is to force metadata collection on this class and it can
be achieved by putting some useless Angular decorator.

https://github.com/angular/angular/issues/16566#issuecomment-338188342